move component-protocol to internal/pkg
