+++ internal/apps/dop/component-protocol/components/all.go
@@ -36,5 +36,5 @@ import (
 	_ "github.com/erda-project/erda/internal/apps/dop/component-protocol/components/scenes-import-record"
 	_ "github.com/erda-project/erda/internal/apps/dop/component-protocol/components/test-dashboard"
 	_ "github.com/erda-project/erda/internal/apps/dop/component-protocol/components/test-report"
-	_ "github.com/erda-project/erda/providers/component-protocol/issueFilter"
+	_ "github.com/erda-project/erda/internal/pkg/component-protocol/issueFilter"
 )
