+++ internal/apps/dop/component-protocol/components/project-pipeline-exec-list/myPage/tabsTable/filterContainer/customFilter/condition.go
@@ -22,8 +22,8 @@ import (
 	"github.com/erda-project/erda/apistructs"
 	"github.com/erda-project/erda/internal/apps/dop/component-protocol/components/project-pipeline-exec-list/common"
 	"github.com/erda-project/erda/internal/apps/dop/component-protocol/components/util"
+	"github.com/erda-project/erda/internal/pkg/component-protocol/condition"
 	"github.com/erda-project/erda/pkg/limit_sync_group"
-	"github.com/erda-project/erda/providers/component-protocol/condition"
 )
 
 func (p *CustomFilter) ConditionRetriever() ([]interface{}, error) {
