+++ internal/apps/dop/component-protocol/components/project-runtime/advanceFilter/advanceFilter.go
@@ -32,7 +32,7 @@ import (
 	"github.com/erda-project/erda/bundle"
 	"github.com/erda-project/erda/internal/apps/dop/component-protocol/components/project-runtime/common"
 	"github.com/erda-project/erda/internal/apps/dop/component-protocol/types"
-	"github.com/erda-project/erda/providers/component-protocol/condition"
+	"github.com/erda-project/erda/internal/pkg/component-protocol/condition"
 )
 
 // todo placeholder
