+++ internal/apps/dop/component-protocol/components/requirement-task-overview/common/gshelper/helper.go
@@ -22,7 +22,7 @@ import (
 	"github.com/erda-project/erda-infra/providers/component-protocol/cptype"
 	"github.com/erda-project/erda/apistructs"
 	"github.com/erda-project/erda/internal/apps/dop/dao"
-	"github.com/erda-project/erda/providers/component-protocol/issueFilter"
+	"github.com/erda-project/erda/internal/pkg/component-protocol/issueFilter"
 )
 
 type GSHelper struct {
