+++ internal/pkg/component-protocol/condition/input.go
