+++ internal/pkg/component-protocol/issueFilter/gshelper/helper.go
